add test to envoy control plane
