+++ edge/pkg/envoycontrolplane/dao/envoy_cluster.go
@@ -6,6 +6,7 @@ import (
 )
 
 type Cluster struct {
+	ID    string `orm:"column(id); size(64); pk"`
 	Name  string `orm:"column(name);null;type(text)";pk`
 	Value string `orm:"column(Value);null;type(text)"`
 }
