+++ edge/pkg/envoycontrolplane/dao/envoy_listener.go
@@ -6,6 +6,7 @@ import (
 )
 
 type Listener struct {
+	ID    string `orm:"column(id); size(64); pk"`
 	Name  string `orm:"column(name);null;type(text)";pk`
 	Value string `orm:"column(Value);null;type(text)"`
 }
@@ -30,7 +31,7 @@ func UpdateListener(listener *Listener) error {
 }
 
 func InsertOrUpdateListener(listener *Listener) error {
-	_, err := dbm.DBAccess.Raw("INSERT OR REPLACE INTO cluster (name, value) VALUES (?,?)", listener.Name, listener.Value).Exec()
+	_, err := dbm.DBAccess.Raw("INSERT OR REPLACE INTO listener (name, value) VALUES (?,?)", listener.Name, listener.Value).Exec()
 	klog.V(4).Infof("update result %v", err)
 	return err
 }
