+++ edge/pkg/envoycontrolplane/dao/envoy_router.go
@@ -6,6 +6,7 @@ import (
 )
 
 type Router struct {
+	ID    string `orm:"column(id); size(64); pk"`
 	Name  string `orm:"column(name);null;type(text)";pk`
 	Value string `orm:"column(Value);null;type(text)"`
 }
@@ -30,7 +31,7 @@ func UpdateRouter(router *Router) error {
 }
 
 func InsertOrUpdateRouter(router *Router) error {
-	_, err := dbm.DBAccess.Raw("INSERT OR REPLACE INTO cluster (name, value) VALUES (?,?)", router.Name, router.Value).Exec()
+	_, err := dbm.DBAccess.Raw("INSERT OR REPLACE INTO route (name, value) VALUES (?,?)", router.Name, router.Value).Exec()
 	klog.V(4).Infof("update result %v", err)
 	return err
 }
